Refactor editor Program into CrashReporting, UiContentUpdate and UiContentDrawer
